Revert "msm: mdss: fb: fix panel reconfiguration check for format change"
* commit from caf LA.UM.8.2.r1-06300-sdm660.0
* this lead kernel panic in offline charging mode

[    7.920528] timeout (0) enabling timegen on ctl=0
[    7.920562] ------------[ cut here ]------------
[    7.920564] WARNING: at ../drivers/video/fbdev/msm/mdss_mdp_intf_video.c:1729
[    7.920567] Modules linked in:
[    7.920568]
[    7.920574] CPU: 0 PID: 556 Comm: charger Not tainted 4.4.153-perf+ #1
[    7.920576] Hardware name: Qualcomm Technologies, Inc. SDM 636 PM660 + PM660L MTP E7T (DT)
[    7.920579] task: ffffffcfeab8b800 task.stack: ffffffcfeabb8000
[    7.920594] PC is at mdss_mdp_video_display+0x2c8/0x3f8
[    7.920597] LR is at mdss_mdp_video_display+0x2c8/0x3f8
[    7.920599] pc : [<ffffff83fac2e050>] lr : [<ffffff83fac2e050>] pstate: 60400145
[    7.920601] sp : ffffffcfeabbb740
[    7.920604] x29: ffffffcfeabbb740 x28: 0000000000000000
[    7.920607] x27: ffffffcfed8be018 x26: 000000000000022c
[    7.920610] x25: 0000000040022040 x24: ffffffcfed8cc068
[    7.920613] x23: 0000000000000001 x22: ffffff83fb89a000
[    7.920616] x21: ffffffcfedff11a0 x20: ffffffcfedff11c8
[    7.920618] x19: ffffffcfed8be018 x18: 0000000000000000
[    7.920621] x17: 0000000000000000 x16: 0000000000000000
[    7.920624] x15: 0000000000000000 x14: ffffff83fca14e10
[    7.920627] x13: 000000000000022c x12: 00000001d818255e
[    7.920630] x11: 0000000001312d00 x10: 0000000000000a20
[    7.920632] x9 : ffffffcfeabbb550 x8 : ffffffcff128af63
[    7.920635] x7 : 0000000000000000 x6 : 0000000000000004
[    7.920638] x5 : 0000000003b9897a x4 : 0000000000000000
[    7.920641] x3 : 0000000000000140 x2 : a9c9037c10596dba
[    7.920643] x1 : a9c9037c10596dba x0 : 0000000000000025
[    7.920646]
[    7.920646] PC: 0xffffff83fac2e010:
[    7.920657] e010  9100a2b4 97f3b099 5290d3e1 b27ff3e2 33103ee1 6b01001f d2800141 aa1403e0
[    7.920666] e030  9a819041 942ac583 350000e0 b9400262 b0008d20 52800001 910b9c00 97f53e83
[    7.920675] e050  d4210000 52800020 d2800002 390076a0 52800081 aa1303e0 52800003 97ff8688
[    7.920684] e070  2a0003e2 340000c0 b9401a61 b0008d20 910c4000 97f53e75 d4210000 aa1303e0
[    7.920685]
[    7.920685] LR: 0xffffff83fac2e010:
[    7.920695] e010  9100a2b4 97f3b099 5290d3e1 b27ff3e2 33103ee1 6b01001f d2800141 aa1403e0
[    7.920703] e030  9a819041 942ac583 350000e0 b9400262 b0008d20 52800001 910b9c00 97f53e83
[    7.920712] e050  d4210000 52800020 d2800002 390076a0 52800081 aa1303e0 52800003 97ff8688
[    7.920721] e070  2a0003e2 340000c0 b9401a61 b0008d20 910c4000 97f53e75 d4210000 aa1303e0
[    7.920723]
[    7.920723] SP: 0xffffffcfeabbb700:
[    7.920732] b700  fac2e050 ffffff83 eabbb740 ffffffcf fac2e050 ffffff83 60400145 00000000
[    7.920740] b720  00000000 00000000 94b10400 ffffffcf 00000000 00000080 00000000 00000000
[    7.920751] b740  eabbb790 ffffffcf fac12fb0 ffffff83 00000000 00000000 00000000 00000000
[    7.920759] b760  00000000 00000000 fccad000 ffffff83 fb898000 ffffff83 eabbb918 ffffffcf
[    7.920761]
[    7.920763] ---[ end trace 0ae23b418963385a ]---
[    7.920765] Call trace:
[    7.920770] Exception stack(0xffffffcfeabbb550 to 0xffffffcfeabbb680)
[    7.920776] b540:                                   ffffffcfed8be018 0000008000000000
[    7.920779] b560: 00000000424bd000 ffffff83fac2e050 0000000060400145 ffffff83fca27b98
[    7.920783] b580: ffffffcfeabbb5a0 ffffff83fa909c8c ffffff83fccc1000 ffffff83fccc1c28
[    7.920787] b5a0: ffffffcfeabbb650 ffffff83fa909ec8 ffffff83fca06000 ffffff83fa909e94
[    7.920790] b5c0: ffffffcfedff11a0 ffffff83fb89a000 0000000000000001 ffffffcfed8cc068
[    7.920794] b5e0: 0000000040022040 000000000000022c ffffffcfed8be018 0000000000000000
[    7.920797] b600: ffffffcfeabbb690 a9c9037c10596dba 0000000000000025 a9c9037c10596dba
[    7.920800] b620: a9c9037c10596dba 0000000000000140 0000000000000000 0000000003b9897a
[    7.920804] b640: 0000000000000004 0000000000000000 ffffffcff128af63 ffffffcfeabbb550
[    7.920807] b660: 0000000000000a20 0000000001312d00 00000001d818255e 000000000000022c
[    7.920812] [<ffffff83fac2e050>] mdss_mdp_video_display+0x2c8/0x3f8
[    7.920821] [<ffffff83fac12fb0>] mdss_mdp_display_commit+0xfcc/0x120c
[    7.920826] [<ffffff83fac41a2c>] mdss_mdp_overlay_kickoff+0x1060/0x18d4
[    7.920829] [<ffffff83fac42dd0>] mdss_mdp_overlay_on+0x190/0x1f0
[    7.920835] [<ffffff83fac9946c>] mdss_fb_blank_unblank+0x128/0x264
[    7.920840] [<ffffff83fac99708>] mdss_fb_blank_sub.isra.73+0x160/0x2e4
[    7.920844] [<ffffff83fac9a5f0>] mdss_fb_set_par+0x40c/0x488
[    7.920850] [<ffffff83fabf76fc>] fb_set_var+0x21c/0x30c
[    7.920853] [<ffffff83fabf7a48>] do_fb_ioctl+0x1f0/0x5f8
[    7.920857] [<ffffff83fabf7e8c>] fb_ioctl+0x3c/0x4c
[    7.920865] [<ffffff83fa9df4dc>] do_vfs_ioctl+0x700/0x7e8
[    7.920868] [<ffffff83fa9df624>] SyS_ioctl+0x60/0x88
[    7.920875] [<ffffff83fa883370>] el0_svc_naked+0x24/0x28
[    7.920886] mdss_dsi_post_panel_on: ctrl=0000000000000000 ndx=0
[    7.920888] mdss_dsi_post_panel_on:-
[    8.010937] init: Service 'qcom-post-boot' (pid 557) exited with status 1
[    8.021322] mdss_mdp_video_wait4comp: vsync wait timeout 0, fallback to poll mode

This reverts commit 46b20db2ce8fa40fb4e85fe206695d03365a12ca.

+++ drivers/video/fbdev/msm/mdss_fb.c
@@ -4258,7 +4258,7 @@ static int mdss_fb_set_par(struct fb_info *info)
 				mfd->fbi->var.yres) * mfd->fb_page;
 
 	old_format = mdss_grayscale_to_mdp_format(var->grayscale);
-	if (!IS_ERR_VALUE((unsigned long) old_format)) {
+	if (!IS_ERR_VALUE(old_format)) {
 		if (old_format != mfd->panel_info->out_format)
 			mfd->panel_reconfig = true;
 	}
